requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-experimental>=0.0.62
langchain-text-splitters>=0.2.2
langchain-ollama>=0.1.0
ollama>=0.1.8

chromadb>=0.5.3
pypdf>=3.8

sentence-transformers>=3.0.1
Django>=5.0.6
djangorestframework>=3.15.2

celery>=5.4.0
redis>=5.0.4

channels>=4.1.0
channels-redis>=4.2.0

psycopg2-binary>=2.9.9

python-dotenv>=1.0.1
pydantic>=2.8.2
uvicorn>=0.30.1
gunicorn>=22.0.0
django-redis
django-celery-beat
daphne
mongoengine
django-mongoengine
requests
django_celery_results
djangorestframework-simplejwt

